backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
email-validator>=2.1.0

# ── Authentication ─────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4

# ── Redis & Caching ────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0

# ── Background Tasks ───────────────────────────────────────────────────────────
celery[redis]>=5.4.0
flower>=2.0.0

# ── File Storage (MinIO / S3) ──────────────────────────────────────────────────
boto3>=1.34.0
minio>=7.2.0

# ── Email ──────────────────────────────────────────────────────────────────────
fastapi-mail>=1.4.0
jinja2>=3.1.0

# ── HTTP Utilities ─────────────────────────────────────────────────────────────
httpx>=0.27.0
python-multipart>=0.0.9

# ── WebSockets ─────────────────────────────────────────────────────────────────
websockets>=12.0

# ── Web Push (KORAK 1 Prompta 2 / PRD §5.3) ────────────────────────────────────
# pywebpush 1.14.0 je poslednja stabilna verzija (mart 2024); pinovan
# minor da bi se izbeglo lomljenje pri eventualnom 2.0 redizajnu API-ja.
# cryptography je njegova hard-dependency (signing VAPID JWT + AES-128-GCM
# enkripcija push payload-a) — eksplicitno je listamo da bi pip resolver
# birao verziju kompatibilnu sa Python 3.12 (>=43.0.0 ima oficijelnu
# 3.12 support na sve OS-ove uključujući Alpine).
pywebpush>=1.14.0,<2.0.0
cryptography>=43.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pytz>=2024.1
python-dateutil>=2.9.0
orjson>=3.10.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=5.0.0
factory-boy>=3.3.0
